% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{quotes}
\alias{quotes}
\title{Main function to get a random quote.
By default a source is chosen randomly with random topic.}
\usage{
quotes(source = NULL, topic = NULL)
}
\description{
Main function to get a random quote.
By default a source is chosen randomly with random topic.
}
