% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summon_quote.R
\name{summon_quote}
\alias{summon_quote}
\title{Summon quote}
\usage{
summon_quote()
}
\value{
Words of wisdom.
}
\description{
This function prints one of the sentences.
Call it whenever you need help.
As a matter of fact it is a better version of `?`.
}
\examples{
summon_quote()
}
